- restructure component folder structure - rename
