+++ src/components/Menu/MenuItems/CantoneseComponent.js
@@ -7,7 +7,7 @@ import {
   Item,
   Segment,
 } from "semantic-ui-react";
-import { MenuContext } from "../../../context/menu-context";
+import { MenuContext } from "../../../context/menuContext";
 import Media from "react-media";
 const CantoneseComponentComponent = () => {
   const [menu, setMenu] = useContext(MenuContext);
