+++ src/components/Menu/MenuItems/CombinationsComponent.js
@@ -8,7 +8,7 @@ import {
   Item,
   Segment
 } from "semantic-ui-react";
-import { MenuContext } from "../../context/menu-context";
+import { MenuContext } from "../../../context/menuContext";
 import Media from "react-media";
 
 const CombinationsComponent = () => {
