+++ src/components/Menu/MenuItems/RiceComponent.js
@@ -1,16 +1,16 @@
 import React, { useContext, Fragment } from "react";
 import { Segment, Card, Header, Item, Image } from "semantic-ui-react";
-import Media from "react-media";
+import Media from 'react-media';
 
-import { MenuContext } from "../../../context/menu-context";
+import { MenuContext } from "../../../context/menuContext";
 
-const EggFooYoungComponent = () => {
+const RiceComponent = () => {
   const [menu, setMenu] = useContext(MenuContext);
 
-  const getEggFooYoung = () => {
+  const getRice = () => {
     if (menu.dishes) {
-      let EggFooYoung = menu.dishes.eggFooYoung;
-      return Object.values(EggFooYoung);
+      let Rice = menu.dishes.Rice;
+      return Object.values(Rice);
     } else {
       return [];
     }
@@ -22,7 +22,7 @@ const EggFooYoungComponent = () => {
         <Fragment>
           {matches.mobile && (
             <Segment.Group>
-              {getEggFooYoung().map(dish => {
+              {getRice().map(dish => {
                 return (
                   <Segment key={dish.name}>
                     <Card.Content>
@@ -57,7 +57,7 @@ const EggFooYoungComponent = () => {
               }}
               divided
             >
-              {getEggFooYoung().map(dish => {
+              {getRice().map(dish => {
                 return (
                   <Item key={dish.name}>
                     {dish.image && <Image src={dish.image} size="small" />}
@@ -88,4 +88,4 @@ const EggFooYoungComponent = () => {
   );
 };
 
-export default EggFooYoungComponent;
+export default RiceComponent;
