--- src/components/Menu/MenuListComponent.css
@@ -1,5 +1,0 @@
-.accordian {
-  background-color: none;
-  background: none;
-  font-size: 1.1em;
-}
