--- src/components/Navbar/MenuSearchComponent.css
@@ -1,5 +1,0 @@
-.ui.search>.results {
-  overflow-y: auto;
-  max-height: 300px;
-  font-weight: bold;
-}
